+++ src/MasterTool/Models/ChamsMode.cs
@@ -1,0 +1,9 @@
+namespace MasterTool.Models
+{
+    public enum ChamsMode
+    {
+        Solid = 0,
+        CullFront = 1,
+        Outline = 2,
+    }
+}